Reorganize components
