+++ petstore_ecommerce/src/App.js
@@ -1,14 +1,14 @@
 //import logo from './logo.svg';
 //import './App.css';
-import Navbar from './Navbar';
+import Navbar from './components/Navbar';
 
 function App() {
   return (
     <div className="App">
       <header className="App-header">
-        <Navbar/>
       </header>
       <body>
+        <Navbar/>
       </body>
     </div>
   );
